rename batch_handler to handler
