+++ meilisearch-lib/src/tasks/handlers/empty_handler.rs
