+++ meilisearch-lib/src/tasks/handlers/index_resolver_handler.rs
