+++ meilisearch-lib/src/tasks/handlers/snapshot_handler.rs
